+++ modules/ads/genrevibes_ads_appodeal_native/lib/genrevibes_ads_appodeal_native.dart
@@ -1,6 +1,7 @@
 /// Appodeal native ads for GenRevibes apps.
 library;
 
+export 'src/appodeal_native_ad_placeholder.dart';
 export 'src/appodeal_native_ad_style.dart';
 export 'src/appodeal_native_ad_view.dart';
 export 'src/appodeal_native_ads.dart';